refactor: Move spore transfer and leap to examples-core
