+++ examples/core/src/rgbpp/index.ts
@@ -1,1 +1,2 @@
 export * from './xudt';
+export * from './spore';
