+++ examples/core/src/rgbpp/spore.ts
@@ -1,0 +1,105 @@
+import {
+  getSporeTypeScript,
+  Hex,
+  genTransferSporeCkbVirtualTx,
+  genLeapSporeFromBtcToCkbVirtualTx,
+} from '@rgbpp-sdk/ckb';
+import { sendRgbppUtxos } from '@rgbpp-sdk/btc';
+import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
+import { isMainnet, collector, btcAddress, btcDataSource, btcKeyPair, btcService } from '../utils';
+
+export interface SporeTransferParams {
+  sporeRgbppLockArgs: Hex;
+  toBtcAddress: string;
+  sporeTypeArgs: Hex;
+}
+
+export const transferSporeOnBtc = async ({
+  sporeRgbppLockArgs,
+  toBtcAddress,
+  sporeTypeArgs,
+}: SporeTransferParams): Promise<string> => {
+  const sporeTypeBytes = serializeScript({
+    ...getSporeTypeScript(isMainnet),
+    args: sporeTypeArgs,
+  });
+
+  const ckbVirtualTxResult = await genTransferSporeCkbVirtualTx({
+    collector,
+    sporeRgbppLockArgs,
+    sporeTypeBytes,
+    isMainnet,
+  });
+
+  const { commitment, ckbRawTx } = ckbVirtualTxResult;
+
+  // Send BTC tx
+  const psbt = await sendRgbppUtxos({
+    ckbVirtualTx: ckbRawTx,
+    commitment,
+    tos: [toBtcAddress],
+    ckbCollector: collector,
+    from: btcAddress!,
+    source: btcDataSource,
+    feeRate: 30,
+  });
+  psbt.signAllInputs(btcKeyPair);
+  psbt.finalizeAllInputs();
+
+  const btcTx = psbt.extractTransaction();
+  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
+
+  console.log('BTC TxId: ', btcTxId);
+
+  await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
+
+  return btcTxId;
+};
+
+export interface SporeLeapParams {
+  sporeRgbppLockArgs: Hex;
+  toCkbAddress: string;
+  sporeTypeArgs: Hex;
+}
+export const leapSporeToCkb = async ({
+  sporeRgbppLockArgs,
+  toCkbAddress,
+  sporeTypeArgs,
+}: SporeLeapParams): Promise<string> => {
+  const sporeTypeBytes = serializeScript({
+    ...getSporeTypeScript(isMainnet),
+    args: sporeTypeArgs,
+  });
+
+  const ckbVirtualTxResult = await genLeapSporeFromBtcToCkbVirtualTx({
+    collector,
+    sporeRgbppLockArgs,
+    sporeTypeBytes,
+    toCkbAddress,
+    isMainnet,
+  });
+
+  const { commitment, ckbRawTx } = ckbVirtualTxResult;
+
+  // Send BTC tx
+  const psbt = await sendRgbppUtxos({
+    ckbVirtualTx: ckbRawTx,
+    commitment,
+    tos: [btcAddress!],
+    ckbCollector: collector,
+    from: btcAddress!,
+    source: btcDataSource,
+    feeRate: 30,
+  });
+  psbt.signAllInputs(btcKeyPair);
+  psbt.finalizeAllInputs();
+
+  const btcTx = psbt.extractTransaction();
+  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
+
+  console.log('BTC TxId: ', btcTxId);
+
+  await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
+
+  return btcTxId;
+};
