+++ examples/rgbpp/spore/5-leap-spore-to-ckb.ts
@@ -1,55 +1,9 @@
-import { buildRgbppLockArgs, getSporeTypeScript, Hex, genLeapSporeFromBtcToCkbVirtualTx } from '@rgbpp-sdk/ckb';
-import { sendRgbppUtxos } from '@rgbpp-sdk/btc';
-import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
-import { isMainnet, collector, btcAddress, btcDataSource, btcKeyPair, btcService } from 'examples-core';
+import { buildRgbppLockArgs } from '@rgbpp-sdk/ckb';
+import { leapSporeToCkb, btcService, SporeLeapParams } from 'examples-core';
 
-const transferSpore = async ({
-  sporeRgbppLockArgs,
-  toCkbAddress,
-  sporeTypeArgs,
-}: {
-  sporeRgbppLockArgs: Hex;
-  toCkbAddress: string;
-  sporeTypeArgs: Hex;
-}) => {
-  // The spore type script is from 3-create-spore.ts, you can find it from the ckb tx spore output cells
-  const sporeTypeBytes = serializeScript({
-    ...getSporeTypeScript(isMainnet),
-    args: sporeTypeArgs,
-  });
-
-  const ckbVirtualTxResult = await genLeapSporeFromBtcToCkbVirtualTx({
-    collector,
-    sporeRgbppLockArgs,
-    sporeTypeBytes,
-    toCkbAddress,
-    isMainnet,
-  });
-
-  const { commitment, ckbRawTx } = ckbVirtualTxResult;
-
-  // console.log(JSON.stringify(ckbRawTx))
-
-  // Send BTC tx
-  const psbt = await sendRgbppUtxos({
-    ckbVirtualTx: ckbRawTx,
-    commitment,
-    tos: [btcAddress!],
-    ckbCollector: collector,
-    from: btcAddress!,
-    source: btcDataSource,
-    feeRate: 30,
-  });
-  psbt.signAllInputs(btcKeyPair);
-  psbt.finalizeAllInputs();
-
-  const btcTx = psbt.extractTransaction();
-  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
-
-  console.log('BTC TxId: ', btcTxId);
-
+const leapSpore = async (params: SporeLeapParams) => {
+  const btcTxId = await leapSporeToCkb(params);
   try {
-    await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
     const interval = setInterval(async () => {
       const { state, failedReason } = await btcService.getRgbppTransactionState(btcTxId);
       console.log('state', state);
@@ -70,7 +24,7 @@ const transferSpore = async ({
 
 // Use your real BTC UTXO information on the BTC Testnet
 // rgbppLockArgs: outIndexU32 + btcTxId
-transferSpore({
+leapSpore({
   // The spore rgbpp lock args is from 3-create-spore.ts
   sporeRgbppLockArgs: buildRgbppLockArgs(3, 'd8a31796fbd42c546f6b22014b9b82b16586ce1df81b0e7ca9a552cdc492a0af'),
   toCkbAddress: 'ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsq0e4xk4rmg5jdkn8aams492a7jlg73ue0gc0ddfj',
