+++ examples/rgbpp/spore/launch/1-prepare-cluster.ts
