+++ examples/rgbpp/spore/launch/3-create-spores.ts
