+++ examples/rgbpp/tsconfig.json
@@ -21,6 +21,6 @@
     "skipLibCheck": true,
     "strict": true
   },
-  "include": ["queue/**/*.ts", "xudt/**/*.ts", "core/*.ts"],
+  "include": ["queue/**/*.ts", "xudt/**/*.ts"],
   "exclude": ["node_modules"]
 }
