+++ examples/rgbpp/xudt/local/3-btc-leap-ckb.ts
@@ -19,7 +19,7 @@ import {
   btcService,
 } from 'examples-core';
 
-const jumpFromBtcToCkb = async ({ rgbppLockArgsList, toCkbAddress, xudtTypeArgs, transferAmount }: LeapToCkbParams) => {
+const leapFromBtcToCkb = async ({ rgbppLockArgsList, toCkbAddress, xudtTypeArgs, transferAmount }: LeapToCkbParams) => {
   const xudtType: CKBComponents.Script = {
     ...getXudtTypeScript(isMainnet),
     args: xudtTypeArgs,
@@ -82,7 +82,7 @@ const jumpFromBtcToCkb = async ({ rgbppLockArgsList, toCkbAddress, xudtTypeArgs,
 };
 
 // rgbppLockArgs: outIndexU32 + btcTxId
-jumpFromBtcToCkb({
+leapFromBtcToCkb({
   rgbppLockArgsList: [buildRgbppLockArgs(1, '24e622419156dd3a277a90bcbb40c7117462a18d5329dd1ada320ca8bdfba715')],
   toCkbAddress: 'ckt1qrfrwcdnvssswdwpn3s9v8fp87emat306ctjwsm3nmlkjg8qyza2cqgqq9kxr7vy7yknezj0vj0xptx6thk6pwyr0sxamv6q',
   xudtTypeArgs: '0x1ba116c119d1cfd98a53e9d1a615cf2af2bb87d95515c9d217d367054cfc696b',
